src/main/java/demo/naked/appointment/java8/behaviorParameterization/AppleInventoryUtil.java
package demo.naked.appointment.java8.behaviorParameterization;

import java.util.ArrayList;
import java.util.List;

/**
 * Created by tm1c14 on 17/08/2015.
 */
public class AppleInventoryUtil {

    public static List<Apple> filterApples(List<Apple> inventory, ApplePredicate p) {
        List<Apple> result = new ArrayList<>();
        for(Apple apple : inventory) {
            if(p.test(apple)) {
                result.add(apple);
            }
        }

        return result;
    }

    public static void prettyPrintApple(List<Apple> inventory, AppleFormatter formatter) {
        for(Apple apple: inventory) {
            String output = formatter.accept(apple);
            System.out.println(output);
        }
    }

    public static <T> List<T> filter(List<T> list, Predicate<T> p) {
        List<T> result = new ArrayList<>();
        for(T e: list) {
            if(p.test(e)) {
                result.add(e);
            }
        }

        return result;
    }
}
